Test code rearranged to make discovery work.
Now you can run:

> python3 -m unittest discover

to run all the tests, using discovery, or

> ./test-open-iscsi -l

to list the tests.
