+++ test/.gitignore
@@ -1,2 +1,3 @@
 __pycache__
-*~
+\#*#
+.#*
