+++ test/harness/__init__.py
@@ -2,4 +2,6 @@
 Harness functions for the open-iscsi test suite.
 """
 
-__version__ = "1.0"
+__version__ = '1.0'
+
+__all__ = ['util', 'iscsi', 'tests']
